move preferences file
