+++ app/src/main/java/com/dietin/data/preferences/OnboardingManager.kt
@@ -1,19 +1,19 @@
-package com.dietin.ui.screen.onboarding
+package com.dietin.data.preferences
 
 import android.content.Context
 import android.content.SharedPreferences
 
 class OnboardingManager(context: Context) {
-    private val sharedPreferences: SharedPreferences =
+    private val onBoardingPreferences: SharedPreferences =
         context.getSharedPreferences("isOnBoardingCompleted", Context.MODE_PRIVATE)
 
     fun saveData(key: String, value: Boolean) {
-        val editor = sharedPreferences.edit()
+        val editor = onBoardingPreferences.edit()
         editor.putBoolean(key, value)
         editor.apply()
     }
 
     fun getData(key: String, defaultValue: Boolean): Boolean {
-        return sharedPreferences.getBoolean(key, defaultValue)
+        return onBoardingPreferences.getBoolean(key, defaultValue)
     }
 }
